+++ start/src/App.vue
@@ -7,7 +7,7 @@
 <template>
   <f-button type="primary" @click="visible1 = true">打开</f-button>
 
-  <f-dialog v-model:visible="visible1" title="标题文字" :mask-close="false">
+  <f-dialog v-model:visible="visible1" title="标题文字" :mask-close="false" mask-background="red" :z-index="5000" :mask-opacity="1">
     这是一个 Dialog
 
     <template #footer>